requirements.txt
albumentations==1.3.0
Flask==1.1.1
imageio==2.22.4
ipython==8.4.0
matplotlib==3.5.3
nibabel==4.0.2
nilearn==0.9.2
numpy==1.23.1
pandas==1.4.3
Pillow==9.3.0
scikit_image==0.19.3
scikit_learn==1.1.2
scipy==1.9.1
seaborn==0.12.2
torch==1.12.1
tqdm==4.64.1